+++ forge/src/main/java/net/gazebo/forge/PlatformImpl.java
@@ -1,10 +1,10 @@
-package net.gazebo.neoforge;
+package net.gazebo.forge;
 
 import net.gazebo.Platform;
-import net.neoforged.fml.loading.LoadingModList;
+import net.minecraftforge.fml.loading.LoadingModList;
 
 public class PlatformImpl {
-    public static class NeoForgeUtil implements Platform.Util {
+    public static class ForgeUtil implements Platform.Util {
         @Override
         public boolean isModLoaded(String modid) {
             // Exact same check as SpellEngine's Platform.Util: LoadingModList (not ModList) is populated
@@ -12,7 +12,7 @@ public class PlatformImpl {
             return LoadingModList.get().getModFileById(modid) != null;
         }
     }
-    private static final Platform.Util UTIL = new NeoForgeUtil();
+    private static final Platform.Util UTIL = new ForgeUtil();
     public static Platform.Util util() {
         return UTIL;
     }